requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
httpx==0.28.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
